sinden mame borders

+++ package/batocera/core/batocera-configgen/configgen/configgen/generators/mame/mameGenerator.py
@@ -19,6 +19,7 @@ from PIL import Image, ImageOps
 from . import mameControllers
 from pathlib import Path
 import csv
+import controllersConfig
 
 eslog = get_logger(__name__)
 
@@ -401,11 +402,11 @@ class MameGenerator(Generator):
             bezelSet = None
         try:
             if messMode != -1:
-                MameGenerator.writeBezelConfig(bezelSet, system, rom, messSysName[messMode])
+                MameGenerator.writeBezelConfig(bezelSet, system, rom, messSysName[messMode], gameResolution, controllersConfig.gunsNeedBorders(guns))
             else:
-                MameGenerator.writeBezelConfig(bezelSet, system, rom, "")
+                MameGenerator.writeBezelConfig(bezelSet, system, rom, "", gameResolution, controllersConfig.gunsNeedBorders(guns))
         except:
-            MameGenerator.writeBezelConfig(None, system, rom, "")
+            MameGenerator.writeBezelConfig(None, system, rom, "", gameResolution, controllersConfig.gunsNeedBorders(guns))
 
         buttonLayout = getMameControlScheme(system, romBasename)
 
@@ -451,7 +452,7 @@ class MameGenerator(Generator):
             old.unlink()
 
     @staticmethod
-    def writeBezelConfig(bezelSet, system, rom, messSys):
+    def writeBezelConfig(bezelSet, system, rom, messSys, gameResolution, gunsNeedBorders):
         romBase = os.path.splitext(os.path.basename(rom))[0] # filename without extension
 
         if messSys == "":
@@ -462,19 +463,32 @@ class MameGenerator(Generator):
         if os.path.exists(tmpZipDir):
             shutil.rmtree(tmpZipDir)
 
-        if bezelSet is None:
+        if bezelSet is None and not gunsNeedBorders:
             return
 
         # let's generate the zip file
         os.makedirs(tmpZipDir)
 
         # bezels infos
-        bz_infos = bezelsUtil.getBezelInfos(rom, bezelSet, system.name, 'mame')
+        if bezelSet is None:
+            if gunsNeedBorders:
+                bz_infos = None
+            else:
+                return
+        else:
+            bz_infos = bezelsUtil.getBezelInfos(rom, bezelSet, system.name, 'mame')
+            if bz_infos is None:
+                if not gunsNeedBorders:
+                    return
+
+        # create an empty bezel
         if bz_infos is None:
-            return
+            overlay_png_file = "/tmp/bezel_transmame_black.png"
+            bezelsUtil.createTransparentBezel(overlay_png_file, gameResolution["width"], gameResolution["height"])
+            bz_infos = { "png": overlay_png_file }
 
         # copy the png inside
-        if os.path.exists(bz_infos["mamezip"]):
+        if "mamezip" in bz_infos and os.path.exists(bz_infos["mamezip"]):
             if messSys == "":
                 artFile = "/var/run/mame_artwork/" + romBase + ".zip"
             else:
@@ -485,15 +499,16 @@ class MameGenerator(Generator):
                 else:
                     os.remove(artFile)
             os.symlink(bz_infos["mamezip"], artFile)
+            # hum, not nice if guns need borders
             return
-        elif os.path.exists(bz_infos["layout"]):
+        elif "layout" in bz_infos and os.path.exists(bz_infos["layout"]):
             os.symlink(bz_infos["layout"], tmpZipDir + "/default.lay")
             pngFile = os.path.split(bz_infos["png"])[1]
             os.symlink(bz_infos["png"], tmpZipDir + "/" + pngFile)
         else:
             pngFile = "default.png"
             os.symlink(bz_infos["png"], tmpZipDir + "/default.png")
-            if os.path.exists(bz_infos["info"]):
+            if "info" in bz_infos and os.path.exists(bz_infos["info"]):
                 bzInfoFile = open(bz_infos["info"], "r")
                 bzInfoText = bzInfoFile.readlines()
                 bz_alpha = 1.0 # Just in case it's not set in the info file
@@ -590,6 +605,16 @@ class MameGenerator(Generator):
             imgnew.paste(back, (0,0,img_width,img_height))
             imgnew.save(output_png_file, mode="RGBA", format="PNG")
 
+            try:
+                os.remove(tmpZipDir + "/" + pngFile)
+            except:
+                pass
+            os.symlink(output_png_file, tmpZipDir + "/" + pngFile)
+
+        # borders for guns
+        if gunsNeedBorders:
+            output_png_file = "/tmp/bezel_gunborders.png"
+            borderSize = bezelsUtil.gunBorderImage(tmpZipDir + "/" + pngFile, output_png_file)
             try:
                 os.remove(tmpZipDir + "/" + pngFile)
             except:
